vera-api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]>=0.29.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.29
psycopg2-binary>=2.9.6
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0

# LLM Providers
openai==1.53.0
groq==0.5.0

# Voice/Audio
elevenlabs==1.50.3

# Utilities
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
aiohttp==3.9.5
